Ajustes varios

+++ Services/Security/Security.Core/Entities/Base/BaseEntity.cs
@@ -8,7 +8,7 @@ namespace Security.Core.Entities.Base
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public Int64 Id { get; set; }
+        public Int32 Id { get; set; }
         //public DateTime CreatedDate { get; set; }
         //public DateTime ModifiedDate { get; private set; }
